modified for S908E

+++ unica/patches/bt-lib-patch/customize.sh
@@ -1,9 +1,16 @@
+# ==============================================================================
+#
+# MOD_NAME="Bluetooth library patcher"
+# MOD_AUTHOR="3arthur6 & duhansysl "
+# MOD_DESC="Fixes Bluetooth JNI issues for ArtisanROM source (skip missing sequences)."
+#
+# ==============================================================================
+
+# Extract libbluetooth_jni.so if missing
 if [ ! -f "$WORK_DIR/system/system/lib64/libbluetooth_jni.so" ]; then
     LOG_STEP_IN "- Extracting libbluetooth_jni.so from com.android.bt.apex"
 
-    if [ -d "$TMP_DIR" ]; then
-        EVAL "rm -rf \"$TMP_DIR\""
-    fi
+    [ -d "$TMP_DIR" ] && EVAL "rm -rf \"$TMP_DIR\""
     mkdir -p "$TMP_DIR"
 
     EVAL "unzip -j \"$WORK_DIR/system/system/apex/com.android.bt.apex\" \"apex_payload.img\" -d \"$TMP_DIR\""
@@ -27,8 +34,20 @@ if [ ! -f "$WORK_DIR/system/system/lib64/libbluetooth_jni.so" ]; then
     LOG_STEP_OUT
 fi
 
-# Disable VaultKeeper support
-# Before: [tbnz w8, #0, #0xbd260]
-# After: [b #0xbd260]
-HEX_PATCH "$WORK_DIR/system/system/lib64/libbluetooth_jni.so" \
-    "2897773948050037" "289777392a000014"
+# Apply patches for SDK 36
+if [ "$SOURCE_PLATFORM_SDK_VERSION" -eq 36 ]; then
+    # Disable VaultKeeper support (S22 Ultra base)
+    # Before: 00122a0140395f01086b00020054
+    # After:  00122a0140395f01086bde030014
+    HEX_PATCH "$WORK_DIR/system/system/lib64/libbluetooth_jni.so" \
+        "00122a0140395f01086b00020054" "00122a0140395f01086bde030014" || LOG "- Skipping missing S22 Ultra patch"
+
+    # Disable VaultKeeper support (S22 base)
+    # Before: 2897773948050037
+    # After:  289777392a000014
+    HEX_PATCH "$WORK_DIR/system/system/lib64/libbluetooth_jni.so" \
+        "2897773948050037" "289777392a000014" || LOG "- Skipping missing S22 base patch"
+
+else
+    LOG "\033[0;33m! Unsupported SDK/API level: $SOURCE_PLATFORM_SDK_VERSION — skipping Bluetooth patch\033[0m"
+fi
